requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
azure-ai-projects
azure-identity
python-multipart
openai
httpx
pytest
mcp
httpx
uvicorn
starlette